Switch from hello-kubecon to any-charm
